Remove extra unit from enum constructors. Remove redundant deregistration of callbacks. Rename private -> extfn.
